# Conflicts:
#	PROJECT-Rogue/Assets/Graphics/Sprites/Water.png.meta

+++ PROJECT-Rogue/Assets/Scripts/Characters/Enemy.cs
@@ -32,9 +32,9 @@ public abstract class Enemy : FightingCharacter
 
     public abstract void rotate();
 
-    public override void TakeDamage(int damage)
+    public override void TakeDamage(float damage)
     {
-        healthPoints -= damage;
+        healthPoints -= (int)damage;
         if (healthPoints <= 0)
         {
             Level.CheckStatus();
