+++ PROJECT-Rogue/Assets/Scripts/Items/Item.cs
@@ -9,7 +9,7 @@ public abstract class Item : MonoBehaviour
 
     protected int maxHPIncrease = 0;
     protected float attackSpeedModification = 0.0f;
-    protected int damageModification = 0;
+    protected float damageModification = 0;
     protected float moveSpeedModification = 0.0f;
     protected float shotSpeedModification = 0.0f;
     protected float rangeModification = 0.0f;
